Primary Component Ready with AuthState & Child App
