+++ src/index.tsx
@@ -1,34 +1,25 @@
 import React from 'react';
-import {Text, View, StyleSheet} from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import type {NavigationContainerProps} from '@react-navigation/native';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
 
-export const Txt = () => {
+//Screens
+import LoginScreen from './screens/Login';
+interface IAuthScreen {
+  isAuthenticated?: boolean;
+  children?: React.ReactNode;
+}
+
+const Stack = createNativeStackNavigator();
+
+export default ({isAuthenticated, children}: IAuthScreen) => {
   return (
-    <View style={styles.body}>
-      <Text style={styles.text}>Auth Component</Text>
-    </View>
+    <Stack.Navigator>
+      {!isAuthenticated ? (
+        <>
+          <Stack.Screen name="Login" component={LoginScreen} />
+        </>
+      ) : (
+        children
+      )}
+    </Stack.Navigator>
   );
 };
-export default () => {
-  return (
-    <View style={styles.body}>
-      <Text style={styles.text}>Auth Component</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  body: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#ffffff',
-  },
-  text: {
-    fontSize: 18,
-    fontWeight: '700',
-    textAlign: 'center',
-    width: '80%',
-  },
-});
