+++ src/screens/Login.tsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import {Text, View, StyleSheet} from 'react-native';
+
+export default () => {
+  return (
+    <View style={styles.body}>
+      <Text style={styles.text}>Login Screen</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  body: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#ffffff',
+  },
+  text: {
+    fontSize: 18,
+    fontWeight: '700',
+    textAlign: 'center',
+    width: '80%',
+  },
+});
